% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HotellingTSq.R
\name{HotellingTSq}
\alias{HotellingTSq}
\title{Hotelling T Squared Test}
\usage{
HotellingTSq(data, H0, alpha = 0.05)
}
\arguments{
\item{data}{the multivariate input data in matrix or dataframe form to be evaluated}

\item{H0}{the hypothesized mean vector}

\item{alpha}{the alpha value used in the determination of the critical value}
}
\value{
True if the null hypothesis is accepted and false if the null hypothesis is rejected
}
\description{
This function tests a provided dataset against a hypothesized mean vector for conformity.
}
\examples{
vec1 <- rnorm(100)
vec2 <- rnorm(100)
vec3 <- rnorm(100)
data <- data.frame(x=vec1, y=vec2, z=vec3)
H0 <- c(0, 0, 0)
alpha <- 0.05
HotellingTSq(data, H0, alpha)
}
