% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA.R
\name{PCA}
\alias{PCA}
\title{Principle Component Analysis}
\usage{
PCA(data)
}
\arguments{
\item{data}{a multivariate data set in matrix or data frame form}
}
\value{
the coefficients of the principle components, the variances of the components, and the cumulative percentage of total variance
}
\description{
This function performs a principle component analysis of the provided data.
}
\examples{
data <- T8_5
PCA(data)
}
