% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{T4_3}
\alias{T4_3}
\title{Table 4.3 (Four Measurements of Stiffness)}
\format{
a data frame containing 30 rows and 4 columns:
\itemize{
  \item x1: First measure of stiffness
  \item x2: Second measure of stiffness
  \item x3: Third measure of stiffness
  \item x4: Fourth measure of stiffness
}
Note: There were no units or scale given with this data and the 5th column d^2 was removed for analysis as it was not
part of the original data.
}
\source{
Applied Multivariate Statistical Analysis 6th Edition by R. Johnson and D. Wichern
}
\usage{
T4_3
}
\description{
A data set containing information regarding stiffness measurements for 30 observations.
}
\keyword{datasets}
