% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{T5_1}
\alias{T5_1}
\title{Table 5.1 (Sweat Data)}
\format{
a data frame containing 20 rows and 3 columns:
\itemize{
  \item SweatRate: Arbitrary measurement of the rate of excretion
  \item Sodium: Relative level of sodium in sweat
  \item Potassium: Relative level of sodium in sweat
}
Note: There were no units or scale given with this data.
}
\source{
Applied Multivariate Statistical Analysis 6th Edition by R. Johnson and D. Wichern
}
\usage{
T5_1
}
\description{
A data set containing information regarding sweating for 20 individuals
}
\keyword{datasets}
