% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{T7_1}
\alias{T7_1}
\title{Table 7.1 (Real-Estate Data)}
\format{
a data frame containing 20 rows and 3 columns:
\itemize{
  \item Total dwelling size: The square footage of the home in hundreds of square feet
  \item Assessed value: The value of the home in thousands of dollars
  \item Selling price: The selling price of the home in thousands of dollars
}
}
\source{
Applied Multivariate Statistical Analysis 6th Edition by R. Johnson and D. Wichern
}
\usage{
T7_1
}
\description{
A data set containing information regarding homes in a Milwaukee, Wisconsin neighborhood
}
\keyword{datasets}
