% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{T8_5}
\alias{T8_5}
\title{Table 8.5 (Census-tract Data)}
\format{
a data frame containing 61 rows and 5 columns:
\itemize{
  \item Total population: The total population the area in thousands of people
  \item Profession: The percentage of people with professional degrees in the area
  \item Employment (%): The percentage of people over the age of 16 employed in the area
  \item Government Employment (%): The percentage of people employed by the government in the area
  \item Medium Home Value: The median home value for the area in hundreds of thousands of dollars
}
}
\source{
Applied Multivariate Statistical Analysis 6th Edition by R. Johnson and D. Wichern
}
\usage{
T8_5
}
\description{
A data set containing information regarding census demographics
}
\keyword{datasets}
